Extend support for posix_clock_settime, including for Pico
Get it work with platforms that only define settimeofday
Also fix logic of clock_id parameter
Also change detection of symbols on Pico and ESP32:
- on both platforms, do not try to link as try_compile link isn't done the way
  we actually link AtomVM and therefore fails to detect availability of
  clock_settime (esp32) and settimeofday (esp32 and pico)
- on both platforms, override the check for mkfifo as the function is declared
  in newlib header but not available (this is why we tried to link initially)
- on Pico, also override the check for unlink as it always fails (and the sdk
  warns about it)
Also move posix_clock_settime implementation to posix_nifs.c
