+++ src/libAtomVM/nifs.c
@@ -660,11 +660,6 @@ static const struct Nif atomvm_read_priv_nif =
     .base.type = NIFFunctionType,
     .nif_ptr = nif_atomvm_read_priv
 };
-static const struct Nif atomvm_posix_clock_settime_nif =
-{
-    .base.type = NIFFunctionType,
-    .nif_ptr = nif_atomvm_posix_clock_settime
-};
 static const struct Nif console_print_nif =
 {
     .base.type = NIFFunctionType,
@@ -761,6 +756,13 @@ DEFINE_MATH_NIF(tanh)
 #define IF_HAVE_UNLINK(expr) (expr)
 #else
 #define IF_HAVE_UNLINK(expr) NULL
+#endif
+#if HAVE_CLOCK_SETTIME
+#define IF_HAVE_CLOCK_SETTIME_OR_SETTIMEOFDAY(expr) (expr)
+#elif HAVE_SETTIMEOFDAY
+#define IF_HAVE_CLOCK_SETTIME_OR_SETTIMEOFDAY(expr) (expr)
+#else
+#define IF_HAVE_CLOCK_SETTIME_OR_SETTIMEOFDAY(expr) NULL
 #endif
 
 //Ignore warning caused by gperf generated code
@@ -3808,52 +3810,6 @@ static term nif_atomvm_read_priv(Context *ctx, int argc, term argv[])
 
     free(complete_path);
     return result;
-}
-
-static term nif_atomvm_posix_clock_settime(Context *ctx, int argc, term argv[])
-{
-    UNUSED(argc);
-
-    VALIDATE_VALUE(argv[0], term_is_atom);
-    if (globalcontext_is_term_equal_to_atom_string(ctx->global, argv[0], ATOM_STR("\x8", "realtime"))) {
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-    VALIDATE_VALUE(argv[1], term_is_tuple);
-    if (term_get_tuple_arity(argv[1]) != 2) {
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-#ifdef HAVE_CLOCK_SETTIME
-
-    term secs = term_get_tuple_element(argv[1], 0);
-    VALIDATE_VALUE(secs, term_is_any_integer);
-    avm_int64_t s = term_maybe_unbox_int64(secs);
-
-    term nsecs = term_get_tuple_element(argv[1], 1);
-    VALIDATE_VALUE(nsecs, term_is_any_integer);
-    avm_int64_t ns = term_maybe_unbox_int64(nsecs);
-
-    struct timespec tp = {
-        .tv_sec = s,
-        .tv_nsec = ns
-    };
-
-    int res = clock_settime(CLOCK_REALTIME, &tp);
-    if (res != 0) {
-        if (UNLIKELY(memory_ensure_free(ctx, TUPLE_SIZE(2)) != MEMORY_GC_OK)) {
-            RAISE_ERROR(OUT_OF_MEMORY_ATOM);
-        }
-        term error = term_alloc_tuple(2, &ctx->heap);
-        term_put_tuple_element(error, 0, ERROR_ATOM);
-        term_put_tuple_element(error, 1, posix_errno_to_term(errno, ctx->global));
-        return error;
-    } else {
-        return OK_ATOM;
-    }
-#else
-    RAISE_ERROR(UNDEF_ATOM);
-#endif
 }
 
 static term nif_console_print(Context *ctx, int argc, term argv[])
